Refactor handler package, splitting up into separate files according to their responsibilities
